# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Next Clear Look Engine API
  version: 1.0.0
  summary: Sentinel-2 geometric opportunities, AOI archive evidence and historical likelihoods.
  description: |
    HTTP contract between the engine and the web app. An opportunity is geometry, not promised acquisition.
    A likelihood is empirical history, not a weather forecast. All instants are aware UTC RFC 3339.
servers:
  - url: http://localhost:8000/v1
    description: Local engine
...
paths:
...
  /satellites/{satellite_id}:
    parameters:
      - $ref: '#/components/parameters/SatelliteId'
    get:
      tags: [Satellites]
      operationId: getSatellite
      summary: Get current catalogue evidence for one satellite.
      responses:
        '200':
          description: Satellite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
              examples:
                sentinel_2c:
                  externalValue: ./examples/satellite.json
        '404':
          $ref: '#/components/responses/NotFound'
  /satellites/{satellite_id}/trajectory:
    parameters:
      - $ref: '#/components/parameters/SatelliteId'
    get:
      tags: [Satellites]
      operationId: getTrajectory
      summary: Get compact geodetic samples for client-side interpolation.
      parameters:
        - name: start
          in: query
          required: true
          schema: {$ref: '#/components/schemas/Timestamp'}
        - name: end
          in: query
          required: true
          schema: {$ref: '#/components/schemas/Timestamp'}
        - name: step_seconds
          in: query
          schema: {type: integer, minimum: 10, maximum: 600, default: 60}
      responses:
        '200':
          description: Sampled trajectory.
          headers:
            X-Provenance-Id:
              $ref: '#/components/headers/ProvenanceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trajectory'
              examples:
                singapore:
                  externalValue: ./examples/trajectory.json
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /aois/{aoi_id}/opportunities:
    parameters:
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [Opportunities]
      operationId: listOpportunities
      summary: List geometric Sentinel-2 opportunities for an AOI.
      description: >-
        These are geometric opportunities and do not promise acquisition. When from and to are
        omitted, the window starts at the engine clock and ends 14 days later. Supply both values
        to request a different window.
      parameters:
        - name: from
          in: query
          description: Inclusive window start. Must be supplied together with to.
          schema: {$ref: '#/components/schemas/Timestamp'}
        - name: to
          in: query
          description: Exclusive window end. Must be supplied together with from.
          schema: {$ref: '#/components/schemas/Timestamp'}
        - name: satellite_id
          in: query
          schema: {type: string}
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Opportunity page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityPage'
              examples:
                singapore:
                  externalValue: ./examples/opportunities.json
        '202':
          description: Analysis is not cached and has been queued.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                pending:
                  externalValue: ./examples/analysis-job.json
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /aois/{aoi_id}/scenes:
    parameters:
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [Scenes]
      operationId: listScenes
      summary: List intersecting archive scenes and AOI-level statistics when available.
      description: Defaults to the half-open archive window [clock minus 30 days, clock).
      parameters:
        - name: from
          in: query
          schema: {$ref: '#/components/schemas/Timestamp'}
        - name: to
          in: query
          schema: {$ref: '#/components/schemas/Timestamp'}
        - name: minimum_aoi_clear_percent
          in: query
          schema: {type: number, minimum: 0, maximum: 100}
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Scene page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenePage'
              examples:
                singapore:
                  externalValue: ./examples/scenes.json
                empty_archive:
                  externalValue: ./examples/scenes-empty.json
                replay_not_recorded:
                  externalValue: ./examples/scenes-not-recorded.json
        '202':
          description: Archive search is pending; this is distinct from an empty completed archive.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                pending:
                  externalValue: ./examples/analysis-job.json
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /scenes/{scene_id}:
    parameters:
      - $ref: '#/components/parameters/SceneId'
    get:
      tags: [Scenes]
      operationId: getScene
      summary: Get STAC-derived scene metadata and engine resource links.
      responses:
        '200':
          description: Scene.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
              examples:
                singapore:
                  externalValue: ./examples/scene.json
        '404':
          $ref: '#/components/responses/NotFound'
  /scenes/{scene_id}/statistics/{aoi_id}:
    parameters:
      - $ref: '#/components/parameters/SceneId'
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [Scenes]
      operationId: getSceneStatistics
      summary: Get AOI-rasterised SCL statistics and class counts.
      responses:
        '200':
          description: AOI statistics.
          headers:
            X-Provenance-Id:
              $ref: '#/components/headers/ProvenanceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RasterStatistics'
              examples:
                singapore:
                  externalValue: ./examples/scene-statistics.json
        '202':
          description: Raster analysis queued.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                pending:
                  externalValue: ./examples/analysis-job.json
        '404':
          $ref: '#/components/responses/NotFound'
  /scenes/{scene_id}/thumbnails/{aoi_id}:
    parameters:
      - $ref: '#/components/parameters/SceneId'
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [Scenes]
      operationId: getSceneThumbnail
      summary: Get a true-colour PNG clipped to the AOI, transparent outside it.
      responses:
        '200':
          description: Clipped RGBA PNG.
          headers:
            ETag:
              schema: {type: string}
              description: Quoted SHA-256 of the PNG bytes.
            X-Provenance-Id:
              $ref: '#/components/headers/ProvenanceId'
            Cache-Control:
              schema: {type: string, example: 'public, max-age=31536000, immutable'}
          content:
            image/png:
              schema: {type: string, format: binary}
        '202':
          description: Thumbnail generation queued.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                pending:
                  externalValue: ./examples/analysis-job.json
        '404':
          $ref: '#/components/responses/NotFound'
  /scenes/{scene_id}/thumbnails/{aoi_id}/metadata:
    parameters:
      - $ref: '#/components/parameters/SceneId'
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [Scenes]
      operationId: getSceneThumbnailMetadata
      summary: Get thumbnail dimensions, checksum and lineage link.
      responses:
        '200':
          description: Thumbnail metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThumbnailMetadata'
              examples:
                singapore:
                  externalValue: ./examples/thumbnail-metadata.json
        '404':
          $ref: '#/components/responses/NotFound'
  /aois/{aoi_id}/likelihood:
    parameters:
      - $ref: '#/components/parameters/AoiId'
    get:
      tags: [Likelihood]
      operationId: getLikelihood
      summary: Get historical chance of at least one usable look in 7 and 14 days.
      description: Empirical historical likelihood; not weather forecasting or tasking.
      parameters:
        - name: as_of
          in: query
          schema: {$ref: '#/components/schemas/Timestamp'}
          description: Defaults to the active deterministic clock.
      responses:
        '200':
          description: Likelihood with posterior, interval and sample evidence.
          headers:
            X-Provenance-Id:
              $ref: '#/components/headers/ProvenanceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Likelihood'
              examples:
                singapore:
                  externalValue: ./examples/likelihood.json
                insufficient_data:
                  externalValue: ./examples/likelihood-insufficient.json
        '202':
          description: Likelihood analysis queued.
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                pending:
                  externalValue: ./examples/analysis-job.json
        '404':
          $ref: '#/components/responses/NotFound'
  /provenance/{provenance_id}:
    parameters:
      - $ref: '#/components/parameters/ProvenanceId'
    get:
      tags: [Provenance]
      operationId: getProvenance
      summary: Resolve raw inputs, algorithms and lineage for a derived result.
      responses:
        '200':
          description: Directed acyclic provenance graph.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceGraph'
              examples:
                singapore:
                  externalValue: ./examples/provenance.json
        '404':
          $ref: '#/components/responses/NotFound'
...
  /analysis-jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Jobs]
      operationId: getAnalysisJob
      summary: Get durable job state and terminal result/error.
      responses:
        '200':
          description: Job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                completed:
                  externalValue: ./examples/analysis-job-completed.json
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Jobs]
      operationId: cancelAnalysisJob
      summary: Idempotently request cancellation.
      responses:
        '202':
          description: Cancellation requested or job already cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
              examples:
                cancelling:
                  externalValue: ./examples/analysis-job-cancelling.json
        '409':
          description: Job already reached a non-cancelled terminal state.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    AoiId:
      name: aoi_id
      in: path
      required: true
      schema: {type: string, pattern: '^aoi_[A-Za-z0-9_-]+$'}
    SatelliteId:
      name: satellite_id
      in: path
      required: true
      schema: {type: string, enum: [sentinel-2a, sentinel-2b, sentinel-2c]}
    SceneId:
      name: scene_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 200}
    ProvenanceId:
      name: provenance_id
      in: path
      required: true
      schema: {type: string, pattern: '^prv_[A-Za-z0-9_-]+$'}
    JobId:
      name: job_id
      in: path
      required: true
      schema: {type: string, pattern: '^job_[A-Za-z0-9_-]+$'}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 512}
...
  headers:
    ProvenanceId:
      schema: {type: string, pattern: '^prv_[A-Za-z0-9_-]+$'}
      description: Lineage graph for the returned representation.
  responses:
    NotFound:
      description: The addressed resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              externalValue: ./examples/error.json
    ValidationError:
      description: Request syntax was valid but one or more values failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Timestamp:
      type: string
      format: date-time
      examples: ['2026-10-02T17:43:54Z']
...
    Satellite:
      type: object
      additionalProperties: false
      required: [id, name, norad_catalog_id, platform, status, omm_epoch, omm_age_seconds, provenance_id]
      properties:
        id: {type: string}
        name: {type: string}
        norad_catalog_id: {type: integer, minimum: 1}
        platform: {type: string, enum: [sentinel-2a, sentinel-2b, sentinel-2c]}
        status: {type: string, enum: [operational-observed, catalogued, stale, retired]}
        omm_epoch: {$ref: '#/components/schemas/Timestamp'}
        omm_age_seconds: {type: number}
        provenance_id: {type: string}
...
    TrajectorySample:
      type: object
      additionalProperties: false
      required: [time, latitude, longitude, altitude_m, swath_left, swath_right]
      properties:
        time: {$ref: '#/components/schemas/Timestamp'}
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
        altitude_m: {type: number, minimum: 0}
        swath_left: {$ref: '#/components/schemas/Position'}
        swath_right: {$ref: '#/components/schemas/Position'}
    Trajectory:
      type: object
      additionalProperties: false
      required: [satellite_id, frame, sample_interval_seconds, start_time, end_time, samples, algorithm_version, provenance_id]
      properties:
        satellite_id: {type: string}
        frame: {type: string, const: ITRS/WGS84 geodetic}
        sample_interval_seconds: {type: integer, minimum: 1}
        start_time: {$ref: '#/components/schemas/Timestamp'}
        end_time: {$ref: '#/components/schemas/Timestamp'}
        samples: {type: array, minItems: 2, items: {$ref: '#/components/schemas/TrajectorySample'}}
        algorithm_version: {type: string}
        provenance_id: {type: string}
...
    Opportunity:
      type: object
      additionalProperties: false
      required: [id, kind, aoi_id, satellite_id, platform, entry_time, closest_time, exit_time, duration_seconds, direction, satellite_sunlit, aoi_sun_elevation_deg, illumination, element_age_seconds, swath_width_km, swath_footprint, minimum_ground_track_distance_km, closest_subpoint, acquisition_status, caveat, source, provenance_id]
      properties:
        id: {type: string}
        kind: {type: string, const: geometric_opportunity}
        aoi_id: {type: string}
        satellite_id: {type: string}
        platform: {type: string}
        entry_time: {$ref: '#/components/schemas/Timestamp'}
        closest_time: {$ref: '#/components/schemas/Timestamp'}
        exit_time: {$ref: '#/components/schemas/Timestamp'}
        duration_seconds: {type: number, minimum: 0}
        direction: {type: string, enum: [descending, ascending]}
        satellite_sunlit: {type: boolean}
        aoi_sun_elevation_deg: {type: number, minimum: -90, maximum: 90}
        illumination: {type: string, enum: [daylight, low_sun]}
        element_age_seconds: {type: number}
        swath_width_km: {type: number, exclusiveMinimum: 0}
        swath_footprint: {$ref: '#/components/schemas/GeoJsonGeometry'}
        minimum_ground_track_distance_km: {type: number, minimum: 0}
        closest_subpoint: {$ref: '#/components/schemas/Subpoint'}
        acquisition_status: {type: string, enum: [unknown, observed, not_observed]}
        matched_scene_id: {type: [string, 'null']}
        truncated: {type: boolean, default: false}
        caveat: {type: string}
        source: {$ref: '#/components/schemas/OpportunitySource'}
        provenance_id: {type: string}
    OpportunityPage:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Opportunity'}}
        meta: {$ref: '#/components/schemas/OpportunityPageMeta'}
    OpportunityPageMeta:
      type: object
      additionalProperties: false
      required: [count, next_cursor, as_of, window_start, window_end]
      properties:
        count: {type: integer, minimum: 0}
        next_cursor: {type: [string, 'null']}
        as_of: {$ref: '#/components/schemas/Timestamp'}
        window_start:
          allOf: [{$ref: '#/components/schemas/Timestamp'}]
          description: Inclusive start shared with the likelihood horizons for the default request.
        window_end:
          allOf: [{$ref: '#/components/schemas/Timestamp'}]
          description: Exclusive end shared with the 14-day likelihood horizon for the default request.
...
    Scene:
      allOf:
        - $ref: '#/components/schemas/SceneSummary'
        - type: object
          required: [assets]
          properties:
            assets:
              type: object
              required: [scl, visual]
              properties:
                scl: {type: string, format: uri}
                visual: {type: string, format: uri}
      unevaluatedProperties: false
    ScenePage:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/SceneSummary'}}
        meta: {$ref: '#/components/schemas/ScenePageMeta'}
    ClassCount:
      type: object
      additionalProperties: false
      required: [label, pixels]
      properties:
        label: {type: string}
        pixels: {type: integer, minimum: 0}
...
    RasterStatistics:
      type: object
      additionalProperties: false
      required: [scene_id, aoi_id, algorithm_version, clear_percent, valid_coverage_percent, clear_pixels, valid_pixels, inside_aoi_pixels, invalid_pixels, clear_scl_classes, valid_scl_classes, surface_class_policy, class_counts, source_resolution_m, overview_factor, computed_at, provenance_id]
      properties:
        scene_id: {type: string}
        aoi_id: {type: string}
        algorithm_version: {type: string}
        clear_percent: {type: [number, 'null'], minimum: 0, maximum: 100}
        valid_coverage_percent: {type: number, minimum: 0, maximum: 100}
        clear_pixels: {type: integer, minimum: 0}
        valid_pixels: {type: integer, minimum: 0}
        inside_aoi_pixels: {type: integer, minimum: 0}
        invalid_pixels: {type: integer, minimum: 0}
        clear_scl_classes: {type: array, items: {type: integer}, minItems: 1}
        valid_scl_classes: {type: array, items: {type: integer}, minItems: 1}
        surface_class_policy: {$ref: '#/components/schemas/SurfaceClassPolicy'}
        class_counts:
          type: object
          additionalProperties: {$ref: '#/components/schemas/ClassCount'}
        source_resolution_m: {type: number, exclusiveMinimum: 0}
        overview_factor: {type: integer, minimum: 1}
        computed_at: {$ref: '#/components/schemas/Timestamp'}
        provenance_id: {type: string}
    ThumbnailMetadata:
      type: object
      additionalProperties: false
      required: [scene_id, aoi_id, url, media_type, width, height, byte_length, sha256, transparent_outside_aoi, provenance_id]
      properties:
        scene_id: {type: string}
        aoi_id: {type: string}
        url: {type: string, format: uri-reference}
        media_type: {type: string, const: image/png}
        width: {type: integer, minimum: 1}
        height: {type: integer, minimum: 1}
        byte_length: {type: integer, minimum: 1}
        sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        transparent_outside_aoi: {type: boolean}
        provenance_id: {type: string}
...
    ExcludedEvidence:
      type: object
      additionalProperties: false
      required: [opportunity_id, reason]
      properties:
        opportunity_id: {type: string}
        reason: {type: string}
    Likelihood:
      type: object
      additionalProperties: false
      required: [aoi_id, as_of, computed_at, window_start, window_end, history_start, history_end, interpretation, usable_definition, surface_class_policy, season, acquisition_posterior, clear_posterior, horizons, monte_carlo, quality, trial_evidence, excluded_evidence, provenance_id]
      properties:
        aoi_id: {type: string}
        as_of: {$ref: '#/components/schemas/Timestamp'}
        computed_at: {$ref: '#/components/schemas/Timestamp'}
        window_start:
          allOf: [{$ref: '#/components/schemas/Timestamp'}]
          description: Inclusive start for every likelihood horizon.
        window_end:
          allOf: [{$ref: '#/components/schemas/Timestamp'}]
          description: Exclusive end of the 14-day horizon and default opportunity list.
        history_start: {$ref: '#/components/schemas/Timestamp'}
        history_end: {$ref: '#/components/schemas/Timestamp'}
        interpretation: {type: string}
        usable_definition: {type: object, additionalProperties: true}
        surface_class_policy: {$ref: '#/components/schemas/SurfaceClassPolicy'}
        season: {type: object, additionalProperties: true}
        acquisition_posterior: {$ref: '#/components/schemas/BetaPosterior'}
        clear_posterior: {$ref: '#/components/schemas/BetaPosterior'}
        horizons: {type: array, minItems: 2, maxItems: 2, items: {$ref: '#/components/schemas/HorizonLikelihood'}}
        monte_carlo: {type: object, additionalProperties: true}
        quality: {type: string, enum: [seasonal, all-season-fallback, tile-fallback, insufficient-data, demo-only-sparse]}
        trial_evidence: {type: array, items: {$ref: '#/components/schemas/TrialEvidence'}}
        excluded_evidence: {type: array, items: {$ref: '#/components/schemas/ExcludedEvidence'}}
        provenance_id: {type: string}
...
    ProvenanceGraph:
      type: object
      additionalProperties: false
      required: [id, root_artifact_id, created_at, nodes, edges]
      properties:
        id: {type: string}
        root_artifact_id: {type: string}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        nodes: {type: array, items: {$ref: '#/components/schemas/ProvenanceNode'}}
        edges: {type: array, items: {$ref: '#/components/schemas/ProvenanceEdge'}}
    JobState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    JobType:
      type: string
      enum: [full_analysis, orbit_only, archive_refresh, raster_scene, likelihood_only]
    JobStage:
      type: string
      enum: [catalogue, orbit, archive, raster, likelihood, finalise]
...
    JobResult:
      type: object
      additionalProperties: true
      properties:
        opportunity_count: {type: integer, minimum: 0}
        scene_count: {type: integer, minimum: 0}
        likelihood_url: {type: string, format: uri-reference}
    AnalysisJob:
      type: object
      additionalProperties: false
      required: [id, type, state, progress, stage, created_at, result, error, events_url]
      properties:
        id: {type: string}
        type: {$ref: '#/components/schemas/JobType'}
        aoi_id: {type: [string, 'null']}
        scene_id: {type: [string, 'null']}
        state: {$ref: '#/components/schemas/JobState'}
        progress: {type: number, minimum: 0, maximum: 1}
        stage: {$ref: '#/components/schemas/JobStage'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        started_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        finished_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        cancel_requested_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        result:
          oneOf: [{$ref: '#/components/schemas/JobResult'}, {type: 'null'}]
        error:
          oneOf: [{$ref: '#/components/schemas/ErrorObject'}, {type: 'null'}]
        events_url: {type: string, format: uri-reference}
...
    ErrorObject:
      type: object
      additionalProperties: false
      required: [code, message, retryable, details, request_id]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object, additionalProperties: true}
        request_id: {type: string}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorObject'}